docker compose for ingestion

+++ w1-intro/ingest_data.py
@@ -3,7 +3,7 @@
 
 import os
 import argparse
-
+from pathlib import Path
 from time import time
 
 import pandas as pd
@@ -17,18 +17,20 @@ def main(params):
     db = params.db
     table_name = params.table_name
     url = params.url
-    
+    data_dir = params.data_dir
     # the backup files are gzipped, and it's important to keep the correct extension
     # for pandas to be able to open the file
-    if url.endswith('.csv.gz'):
-        csv_name = 'output.csv.gz'
+    
+    csv_name = Path(data_dir) / Path(url).name
+    if not csv_name.exists():
+        os.system(f"wget {url} -O {csv_name}")
     else:
-        csv_name = 'output.csv'
+        print(f"{csv_name} already exists; no download required")
 
-    os.system(f"wget {url} -O {csv_name}")
-
+    # sqlalchemy
     engine = create_engine(f'postgresql://{user}:{password}@{host}:{port}/{db}')
 
+    # specifying iterator and chunksize returns ... an iterator in those chunks
     df_iter = pd.read_csv(csv_name, iterator=True, chunksize=100000)
 
     df = next(df_iter)
@@ -36,16 +38,19 @@ def main(params):
     df.tpep_pickup_datetime = pd.to_datetime(df.tpep_pickup_datetime)
     df.tpep_dropoff_datetime = pd.to_datetime(df.tpep_dropoff_datetime)
 
+    # specify n=0 to get only column names;
+    # Writing only column names to sql basically creates the table with
+    # the inferred schema for us automatically
     df.head(n=0).to_sql(name=table_name, con=engine, if_exists='replace')
 
+    # write the first chunk
     df.to_sql(name=table_name, con=engine, if_exists='append')
-
 
     while True: 
 
         try:
             t_start = time()
-            
+            # get the next chunk
             df = next(df_iter)
 
             df.tpep_pickup_datetime = pd.to_datetime(df.tpep_pickup_datetime)
@@ -71,6 +76,7 @@ if __name__ == '__main__':
     parser.add_argument('--db', required=True, help='database name for postgres')
     parser.add_argument('--table_name', required=True, help='name of the table where we will write the results to')
     parser.add_argument('--url', required=True, help='url of the csv file')
+    parser.add_argument('--data_dir', required=True, help='output directory of downloaded file')
 
     args = parser.parse_args()
